TweetCSpec: rm test when posting empty Tweet
It appears that one of the commits between a8bb78f
and this point broke a test but did not fix it
during the commit process.  There are a number of
commits that address various issues but several of
them are broken.  As such, I don't think that I
can find the exact change.

In any case, it appears this test is incompatible
with a strong_params `ActionController` universe.

The test, as was, flies in the face of how strong
parameters works since `.require` tests by:

>     def require(key)
>       value = self[key]
>       if value.present? || value == false
>         value
>       else

value, here would be `{}` and, thus would always
trigger the else case since `present?` is based on
value being not `blank?`.